New DB Design

+++ Katiba55.API/Controllers/MediasController.cs
@@ -6,25 +6,25 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Katiba55.API.Controllers
 {
-    [Route("api/projects/{projectId}/medias")]
-    public class ProjectMediasController: BaseController
+    [Route("api/medias")]
+    public class MediasController: BaseController
     {
         private readonly ApplicationDbContext _context;
         private readonly IMapper _mapper;
 
-        public ProjectMediasController(ApplicationDbContext context, IMapper mapper)
+        public MediasController(ApplicationDbContext context, IMapper mapper)
         {
             _context = context;
             _mapper = mapper;
         }
 
         [HttpPost("create")]
-        public async Task<IActionResult> CreateAsync(int projectId, CreateProjectMediaDto dto)
+        public async Task<IActionResult> CreateAsync(CreateProjectMediaDto dto)
         {
             var supportedImageExtensions = new[] { ".jpg", ".jpeg", ".png", ".svg", ".gif", ".bmp", ".webp", ".tiff" };
             var supportedVideosExtensions = new[] { ".mp4", ".avi", ".mov", ".wmv", ".flv", ".mkv", ".webm" };
             var fileExtension = Path.GetExtension(dto.Path);
-            var media = new ProjectMedia { ProjectId = projectId };
+            var media = new Media();
 
             if (supportedImageExtensions.Any(ex => ex == fileExtension))
             {
@@ -41,39 +41,40 @@ namespace Katiba55.API.Controllers
 
             _mapper.Map(dto, media);
 
-            _context.ProjectMedias.Add(media);
+            _context.Medias.Add(media);
             await _context.SaveChangesAsync();
 
             return Response(ResultFactory.Ok(media.Id));
         }
 
         [HttpDelete("{mediaId}/delete")]
-        public async Task<IActionResult> DeleteAsync(int projectId, int mediaId)
+        public async Task<IActionResult> DeleteAsync(int mediaId)
         {
-            var media = await _context.ProjectMedias.FirstOrDefaultAsync(m => m.Id == mediaId && m.ProjectId == projectId);
+            var media = await _context.Medias.FirstOrDefaultAsync(m => m.Id == mediaId);
 
             if (media == null)
                 return Response(ResultFactory.NotFound());
 
             try
             {
-                System.IO.File.Delete(media.Path);
+                if(System.IO.File.Exists(media.Path))
+                    System.IO.File.Delete(media.Path);
             }
             catch
             {
                 return Response(ResultFactory.BadRequest(message: "حدث خطأ أثناء حذف الملف. يرجى المحاولة مرة أخرى."));
             }
 
-            _context.ProjectMedias.Remove(media);
+            _context.Medias.Remove(media);
             await _context.SaveChangesAsync();
 
             return Response(ResultFactory.NoContent());
         }
 
-        [HttpGet("getAll")]
-        public async Task<IActionResult> GetAllAsync(int projectId)
+        [HttpGet("getByProjectId")]
+        public async Task<IActionResult> GetByProjectIdAsync([FromQuery] int projectId)
         {
-            var medias = await _context.ProjectMedias
+            var medias = await _context.Medias
                 .Where(m => m.ProjectId == projectId)
                 .ProjectTo<ProjectMediaDto>(_mapper.ConfigurationProvider)
                 .ToListAsync();
